+++ components/ButtonLink.js
@@ -1,10 +1,12 @@
 import Link from 'next/link';
 
 export default function ButtonLink(props) {
-  const { className, link, children } = props;
+  const { className, link, children, styles } = props;
   return (
     <Link href={link}>
-      <a className={className}>{children}</a>
+      <a className={className} style={styles}>
+        {children}
+      </a>
     </Link>
   );
 }