+++ components/TwoColumn.js
@@ -5,26 +5,47 @@ import buttonStyles from '../styles/ButtonLink.module.scss';
 
 export default function TwoColumn(props) {
   // Add rowOrder="row-reverse" prop to the component to reverse its order on desktop
-  const { image, title, content, rowOrder } = props;
+  const { image, title, content, rowOrder, color, bgColor, btnColor, btnBg } =
+    props;
+
+  const styleProps = {
+    btn: {
+      color: btnColor,
+      backgroundColor: btnBg,
+    },
+    container: {
+      color: color,
+      backgroundColor: bgColor,
+    },
+  };
+
   return (
-    <section className={styles.container}>
+    <section className={styles.container} style={styleProps.container}>
       <div className={styles.inner} style={{ flexDirection: rowOrder }}>
         <div className={styles.inner__content}>
-          <h2 className={styles.title}>{title}</h2>
+          <h2 className={styles.title} style={{ color: color }}>
+            {title}
+          </h2>
           <p className={styles.content}>{content}</p>
-          <ButtonLink link="/about-us" className={buttonStyles.btn}>
+          <ButtonLink
+            link="/about-us"
+            className={buttonStyles.btn}
+            styles={styleProps.btn}
+          >
             Learn more
           </ButtonLink>
         </div>
-        <div className={styles.inner__image}>
-          <Image
-            src={image}
-            alt=""
-            className={styles.img}
-            layout="fill"
-            objectFit="cover"
-          />
-        </div>
+        {image && (
+          <div className={styles.inner__image}>
+            <Image
+              src={image}
+              alt=""
+              className={styles.img}
+              layout="fill"
+              objectFit="cover"
+            />
+          </div>
+        )}
       </div>
     </section>
   );