+++ CHANGELOG.md
@@ -16,10 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- add TwoColumn.js component
 - add new footer and newsletter components styling
 
 ### Fixed
 
 - updated Next.js from 10.0.0 to 12.0.10
 - fixed scroll-x added by the layout container, adjust README
+- updated TwoColumn.js component with style props